requirements.txt
torch==1.12.0
torchvision>=0.12.0
einops
pytorch-lightning==1.6.4
torchmetrics==0.6.0
lightning-bolts>=0.5.0
tqdm
wandb
scipy
timm
scikit-learn
pytorch-ignite